Make sure that we succeed in starting a definition before we complete it and emit an error if we fail to start the definition.
ClangASTContext::StartTagDeclarationDefinition(...) was starting definitions for any TagType instances that have TagDecl, but ClangASTContext::CompleteTagDeclarationDefinition(...) was getting the type to a CXXRecordDecl with:

    clang::CXXRecordDecl *cxx_record_decl = qual_type->getAsCXXRecordDecl();

The problem is that getAsCXXRecordDecl() might dig a bit deeper into a type and dig out a different decl, which means we might call ClangASTContext::StartTagDeclarationDefinition(...), but it might not do anything, and then we might call ClangASTContext::CompleteTagDeclarationDefinition(...) and it might try to complete something that didn't have its definition started and this will crash.

This change fixes that, and also makes sure that starting a definition succeeds before any calls to ClangASTContext::CompleteTagDeclarationDefinition().

<rdar://problem/24091798>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@270891 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 179036917d0475f5fd8d073b3d88e9964d2a9c6d)

apple-llvm-split-commit: e83b0fb7afda586545048664b16a64a003c8876c
apple-llvm-split-dir: lldb/