Made templates that have Enumeration values as arguments work correctly.
We were checking for integer types only before this. So I added the ability for CompilerType objects to check for integer and enum types.

Then I searched for places that were using the CompilerType::IsIntegerType(...) function. Many of these places also wanted to be checking for enumeration types as well, so I have fixed those places. These are in the ABI plug-ins where we are figuring out which arguments would go in where in regisers/stack when making a function call, or determining where the return value would live. The real fix for this is to use clang to compiler a CGFunctionInfo and then modify the code to be able to take the IR and a calling convention and have the backend answer the questions correctly for us so we don't need to create a really bad copy of the ABI in each plug-in, but that is beyond the scope of this bug fix.

Also added a test case to ensure this doesn't regress in the future.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@273750 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 67f0bf0aa130022877e31bb6edbeb5be0f362234)

apple-llvm-split-commit: af4a70111755b724c05880d70cb2626f904f00a1
apple-llvm-split-dir: lldb/