[tsan] Treat non-barrier OSAtomic* functions as barriers to avoid false positives
The non-barrier versions of OSAtomic* functions are semantically mo_relaxed, but the two variants (e.g. OSAtomicAdd32 and OSAtomicAdd32Barrier) are actually aliases of each other, and we cannot have different interceptors for them, because they're actually the same function. Thus, we have to stay conservative and treat the non-barrier versions as mo_acq_rel.

Differential Revision: http://reviews.llvm.org/D21733



git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@273890 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 14220476b50e735b6b015c2d4ff25e6a02351048
apple-llvm-split-dir: compiler-rt/