[tsan] dispatch_once interceptor will cause a crash/deadlock when the original dispatch_once is used
Because we use SCOPED_TSAN_INTERCEPTOR in the dispatch_once interceptor, the original dispatch_once can also be sometimes called (when ignores are enabled or when thr->is_inited is false). However the original dispatch_once function doesn’t expect to find “2” in the storage and it will spin forever (but we use “2” to indicate that the initialization is already done, so no waiting is necessary). This patch makes sure we never call the original dispatch_once.

Differential Revision: http://reviews.llvm.org/D21976



git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@274548 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: a446cf7f85d3ef69d108b1b00c7f643e7949e04d
apple-llvm-split-dir: compiler-rt/