Added an "inline-REPL" test case style that allows REPL tests without pexpect.
These test cases simply alternate expressions with (in /// comment blocks)
regexps to be run against the results of those expressions.  They run the LLDB
expression parser in REPL mode.  This eliminates a dependency on pexpect, which
does not provide helpful errors, is harder to debug, and can be unreliable when
the system running the test is under load.

apple-llvm-split-commit: 119fcb317c16ec30c20b63a1838da39302317352
apple-llvm-split-dir: lldb/