Reapply REPL inline test patch with fixes
This change is by Sean Callanan

Added an "inline-REPL" test case style that allows REPL tests
ithout pexpect.

These test cases simply alternate expressions with (in ///
comment blocks) regexps to be run against the results of those expressions.
They run the LLDB xpression parser in REPL mode.  This eliminates a
dependency on pexpect, which does not provide helpful errors, is harder to debug,
and can be unreliable when the system running the test is under load.

apple-llvm-split-commit: 9a6febbfaff7f16ce2ccac2de75c4a37d54f56e3
apple-llvm-split-dir: lldb/