[ThinLTO] Make sure the Context used for the ThinLTO backend has all the appropriate options
An important performance setting on the LLVMContext for LTO is
enableDebugTypeODRUniquing(), this adds an automatic merging of
debug information in the context based on type ids.

Also, the lto::Config includes a diagnostic handler that needs to
be set on the Context, as well as the setDiscardValueNames() setting.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@279532 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: c2cf7e1ef93f512209dc5b7fe1169c21327ff85e
apple-llvm-split-dir: llvm/