Cherry-pick to fix rdar://29229698
When we reach the end of a #include of a header of a local submodule that we
textually included, create an ImportDecl just as we would if we reached a
determine the set of variables to initialize for an imported module.

This should hopefully make the modules selfhost buildbot green again.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@280409 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 2ecff7244bc95a0374d0202e17b484edb71a2ef4
apple-llvm-split-dir: clang/