[asan] Add a new AddressDescription structure, which can describe any type of address.
Summary:
This is useful for inclusion in the Error* structures, to describe an
arbitrary address.

Remove the old struct since it's used only once. This removes one level of
indirection, and moves all *AddressDescription to be one of the recently
introduced structures.

This merges differential revisions: D24131 and D24132

Reviewers: kcc, eugenis, vitalybuka

Subscribers: kubabrecka, llvm-commits

Differential Revision: https://reviews.llvm.org/D24131

git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@281090 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 1b9200784d57c061ba8b1de4c9bf4e48e3012fb1
apple-llvm-split-dir: compiler-rt/