[analyzer] scan-build-py: Remove relative path hack for SATestsBuild.py
Remove the relative path hack in scan-build-py that converts a fully qualified
directory name and a fully qualified file path to a relative path before running
the analyzer on a file.

This hack is not needed: the bad interaction with SATestsBuild.py it was
intended to address is actually the same underlying problem that r280768 fixed.
Further, because the hack would always relativize paths, it caused
SATestBuild.py to be unable to properly line up issues when the build system
changed directory and then built a source file in a child directory but used a
fully-qualified path for the source file.

Differential Revision: https://reviews.llvm.org/D24470

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@281516 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 82d5b51ee7e451c8672c60b4d96a8ba6db6f71e1)

apple-llvm-split-commit: 64146d9d7122f1347d4fe6c76f68daa0bc713a31
apple-llvm-split-dir: clang/