Reapplying r281895 (and follow-up  r281964) after fixing pr30468.
The additional fix is:

When adding debug information to a lowered phi node in mem2reg
check that we have a valid insertion point after the phi for adding
the debug information.

This change addresses the issue in pr30468 where a lowered phi was
added before a catchswitch and no debug information should be added
after the phi in this case.

Differential Revision: https://reviews.llvm.org/D24797

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@282155 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 523c933b4c032777d1485140dcdb8238a39285d3)

apple-llvm-split-commit: 7b202c3120e1bbdd50181d970a2c1c97cde888df
apple-llvm-split-dir: llvm/